get_data_info

get_data_info

Usage

get_data_info(data_list, ...)

Arguments

data_list

a list of prepared data that is generated by data_ready

...

input information for ESS analysis including gca, ald, location, WESS, gamest, domain, select_domain. Those come from the server inputs.

Value

a list containing all the information necessary for ESS estimation.

Examples

if (FALSE) { # \dontrun{
require(embededss)
filePath <- "data/freq_data.xlsx"
imported_data <- read_data(filePath)
data_list <- data_ready(imported_data)
data_information <- get_data_info(data_list, grade = c("M3"), ald = "ALD", location = "Loc_RP60", wess = F, modal = F, threshold = F)

} # }